time in Long

+++ src/main/kotlin/org/pigletsinc/syncplay/synchronization/model/VideoSyncMessage.kt
@@ -2,7 +2,7 @@ package org.pigletsinc.syncplay.synchronization.model
 
 data class VideoSyncMessage(
     val action: String,
-    val time: Double,
+    val time: Long,
     val clientId: String,
     val videoId: String?,
 )
